+++ app/models/early_career_payments/eligibility.rb
@@ -5,6 +5,7 @@ module EarlyCareerPayments
       :employed_as_supply_teacher,
       :has_entire_term_contract,
       :employed_directly,
+      :subject_to_formal_performance_action,
       :subject_to_disciplinary_action,
       :pgitt_or_ugitt_course,
       :eligible_itt_subject,
@@ -46,6 +47,7 @@ module EarlyCareerPayments
     validates :employed_as_supply_teacher, on: [:"supply-teacher", :submit], inclusion: {in: [true, false], message: "Select yes if you are currently employed as a supply teacher"}
     validates :has_entire_term_contract, on: [:"entire-term-contract", :submit], inclusion: {in: [true, false], message: "Select yes if you have a contract to teach at the same school for one term or longer"}, if: :employed_as_supply_teacher?
     validates :employed_directly, on: [:"employed-directly", :submit], inclusion: {in: [true, false], message: "Select yes if you are employed directly by your school"}, if: :employed_as_supply_teacher?
+    validates :subject_to_formal_performance_action, on: [:"formal-performance-action", :submit], inclusion: {in: [true, false], message: "Select yes if you are subject to formal action for poor performance at work"}
     validates :subject_to_disciplinary_action, on: [:"disciplinary-action", :submit], inclusion: {in: [true, false], message: "Select yes if you are subject to disciplinary action"}
     validates :pgitt_or_ugitt_course, on: [:"postgraduate-itt-or-undergraduate-itt-course", :submit], presence: {message: "Select postgraduate if you did a Postgraduate ITT course"}
     validates :eligible_itt_subject, on: [:"eligible-itt-subject", :submit], presence: {message: "Select if you completed your initial teacher training in Chemistry, Mathematics, Modern Foreign Languages, Physics or None of these subjects"}
@@ -60,6 +62,7 @@ module EarlyCareerPayments
       ineligible_nqt_in_academic_year_after_itt? ||
         no_entire_term_contract? ||
         not_employed_directly? ||
+        subject_to_formal_performance_action? ||
         subject_to_disciplinary_action? ||
         not_supported_itt_subject? ||
         not_teaching_now_in_eligible_itt_subject? ||