+++ db/migrate/20210424130733_add_subject_to_formal_performance_action_to_early_career_payments_eligibilities.rb
@@ -1,0 +1,5 @@
+class AddSubjectToFormalPerformanceActionToEarlyCareerPaymentsEligibilities < ActiveRecord::Migration[6.0]
+  def change
+    add_column :early_career_payments_eligibilities, :subject_to_formal_performance_action, :boolean
+  end
+end